% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_obj.R
\name{collapse_tech_repl}
\alias{collapse_tech_repl}
\title{Collapse technical replicates in proteomics data.}
\usage{
collapse_tech_repl(p_df)
}
\arguments{
\item{p_df}{Proteomics data with technical replicates.}
}
\value{
Proteomics data.
}
\description{
Collapse technical replicates in proteomics data.
}
