% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{collapse_uniprot_ids}
\alias{collapse_uniprot_ids}
\title{Collapse UNIPROT IDs}
\usage{
collapse_uniprot_ids(
  uniprot_ids,
  multiple_mode = "first",
  manual_file,
  separator_in = ";",
  separator_out = ";"
)
}
\arguments{
\item{uniprot_ids}{A vector of UNIPROT IDs}

\item{multiple_mode}{Determines how to deal with entries that can not be
broken down into a single entry. Possible values are "none", "manual",
and "first" (default is "first").}

\item{manual_file}{In multiple_mode "manual", the file used for manual
replacements. Required data frame needs to be saved in in a tab separated
file with the columns "input" and "output".}

\item{separator_in}{The separator between multiple UNIPROT IDs and isoforms
in an input (default ";").}

\item{separator_out}{In multiple_mode "none", the separator between
multiple UNIPROT IDs when writing output (default ";").}
}
\value{
Vector of collapes UNIPROT IDs
}
\description{
This function takes a vector of UNIPROT IDs as produced by Maxquant
(in the column "Majority Protein IDs"), collapses isoforms. If after that,
multiple IDs are found, there are different modes to select only one.
}
