% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{construct_contrasts_all}
\alias{construct_contrasts_all}
\title{Construct all pairwise contrasts.}
\usage{
construct_contrasts_all(p_df)
}
\arguments{
\item{p_df}{Proteomics data.}
}
\value{
data frame, the constrasts in the expected format.
}
\description{
Construct all pairwise contrasts.
}
