% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{construct_contrasts_control}
\alias{construct_contrasts_control}
\title{Construct contrasts against control group.}
\usage{
construct_contrasts_control(p_df, cntrl_group)
}
\arguments{
\item{p_df}{Proteomics data.}

\item{cntrl_group}{character, the control group.}
}
\value{
data frame, the contrasts against the control group in the expected format.
}
\description{
Construct contrasts against control group.
}
