% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_obj.R
\name{deconstruct_groups}
\alias{deconstruct_groups}
\title{Deconstruct groups in proteomics data.}
\usage{
deconstruct_groups(p_df)
}
\arguments{
\item{p_df}{Proteomics data.}
}
\value{
TODO
}
\description{
TODO WHAT DOES THIS DO AGAIN?
}
