% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{diff_expr}
\alias{diff_expr}
\title{Differential analysis of protemic data set with different methods.}
\usage{
diff_expr(p_df, method, contrasts, adjust_method = "BH")
}
\arguments{
\item{p_df}{The data to run differential statistical analysis on.
Of class \code{proteomics_data}.}

\item{method}{The method used for differential analysis. See
\code{stat_analysis_methods()} for available options.}

\item{contrasts}{The contrasts between condition to test in differential
analysis. Expected to be a data frame with two columns, a and b, where
each row corresponds to a contrast to test.}

\item{adjust_method}{The method used for correcting for multiple hypothesis
testing. See \code{p.adjust.methods} for available options.}
}
\value{
A data frame with calculated log fold changes and adjusted pvalues
for each specified contrast.
}
\description{
Differential analysis of protemic data set with different methods.
}
