% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{diff_type}
\alias{diff_type}
\title{Assign comparison type.}
\usage{
diff_type(p_df, contrasts)
}
\arguments{
\item{p_df}{Proteomics data.}

\item{contrasts}{data frame, the contrasts for which the type should be determined.
Expected in the same format as for \code{diff_expr}.}
}
\value{
Data frame with the contrast types.
}
\description{
Construct info whether comparisons are based on value-value, value-imput,
or imput-imput, for each protein and each contrasts given.
}
\details{
In order for this comparison to be accurate, the proteomics data from after
filtering, but before imputation should be used.
}
