% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{.filtering_any_group}
\alias{.filtering_any_group}
\title{Filtering data set by presence of values in at least one group in the data
set}
\usage{
.filtering_any_group(data, threshold, annotation)
}
\arguments{
\item{data}{The data frame to run filtering on.}

\item{threshold}{The threshold how many samples need to be present in the
subset the different methods look at to retain a data point.}

\item{annotation}{A data frame that contains information about which samples
in the input data frame belong to which condition.}
}
\value{
A filtered data frame.
}
\description{
Filtering data set by presence of values in at least one group in the data
set
}
