% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{.get_mar_matrix}
\alias{.get_mar_matrix}
\title{Obtain matrix of MAR values.}
\usage{
.get_mar_matrix(mx, groups)
}
\arguments{
\item{mx}{A expression matrix with missing values.}

\item{groups}{A vector indicating which columns belong to the same group.}
}
\value{
A logical matrix of the same dimension as mx. TRUE values are
are considered MAR.
}
\description{
Missing at random values are defined as observations where not all values
for a specific group/condition are missing, but only some.
}
