% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{.imputation_mixed_sample}
\alias{.imputation_mixed_sample}
\title{Sample-wise mixed imputation}
\usage{
.imputation_mixed_sample(mx, mar_mx, mar_method, mnar_method, ...)
}
\arguments{
\item{mx}{A expression matrix with missing values.}

\item{mar_mx}{A logical matrix of the same dimensions as mx. TRUE values
are considered MAR.}

\item{mar_method}{The method used for imputing MAR missing values.}

\item{mnar_method}{The method used for imputing MNAR missing values.}

\item{...}{Other parameters passed on to mar_method and mnar_method.}
}
\value{
A matrix of the same dimensions as the input with imputed values.
}
\description{
Impute expression matrix with a mixed imputation using a sample-wise
separation in MAR and MNAR missing values
}
