% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{.stat_analysis_limma}
\alias{.stat_analysis_limma}
\title{Differential analysis using LIMMA}
\usage{
.stat_analysis_limma(mx, groups, contrasts)
}
\arguments{
\item{mx}{An expression matrix.}

\item{groups}{A vector indicating which columns belong to which group}

\item{contrasts}{A vector of contrasts in limma format.}
}
\value{
A data frame with raw pvalues.
}
\description{
Differential analysis using LIMMA
}
