% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{.stat_analysis_ttest}
\alias{.stat_analysis_ttest}
\title{Differential analysis using T-tests.}
\usage{
.stat_analysis_ttest(mx, groups, contrasts)
}
\arguments{
\item{mx}{An expression matrix.}

\item{groups}{A vector indicating which columns belong to which group}

\item{contrasts}{The contrasts between condition to test in differential
analysis. Expected to be a data frame with two columns, a and b, where
each row corresponds to a contrast to test.}
}
\value{
A data frame with raw pvalues.
}
\description{
Differential analysis using T-tests.
}
