% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_obj.R
\name{filter_data_by_group}
\alias{filter_data_by_group}
\title{Filter proteomics data by groups.}
\usage{
filter_data_by_group(p_df, groups, keep = TRUE)
}
\arguments{
\item{p_df}{Proteomics data.}

\item{groups}{Vector of groups that the data set will be filtered by.}

\item{keep}{Logical, whether members of groups should be kept in the
data set or discarded (default is TRUE, so specified groups will be kept).}
}
\value{
Filtered proteomics data.
}
\description{
Filter proteomics data by groups.
}
