% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_obj.R
\name{filter_data_by_group_and_id}
\alias{filter_data_by_group_and_id}
\title{Filer proteomics data by groups and ids.}
\usage{
filter_data_by_group_and_id(p_df, df_filter)
}
\arguments{
\item{p_df}{Proteomics data.}

\item{df_filter}{Dataframe, expected to have two columns called group and id.}
}
\value{
Filtered proteomics data.
}
\description{
This filtering can be useful to filter a data set against the output of a
differential analysis, for example when removing proteins from a data set that
have also been identified in a negative control.
}
