% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_obj.R
\name{filter_data_by_samples}
\alias{filter_data_by_samples}
\title{Filter proteomics data by samples.}
\usage{
filter_data_by_samples(p_df, samples, keep = TRUE)
}
\arguments{
\item{p_df}{Proteomics data.}

\item{samples}{Vector of samples that the data set will be filered by.}

\item{keep}{logical, whether members of samples should be kept in the
data set or discarded (default is TRUE, so specified samples will be kept).}
}
\value{
Filtered proteomics data.
}
\description{
Filter proteomics data by samples.
}
