% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filtering}
\alias{filtering}
\title{Filtering proteomics data set with different methods}
\usage{
filtering(p_df, method, threshold)
}
\arguments{
\item{p_df}{The data to run filtering on. Of class \code{proteomics_data}.}

\item{method}{The method used for filtering. See \code{filtering_methods()} for
available option.}

\item{threshold}{The threshold how many samples need to be present in the
subset the different methods look at to retain a data point.}
}
\value{
A filtered object of class \code{proteomics_data}.
}
\description{
TODO WRITE DETAILS
}
