% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filtering_methods}
\alias{filtering_methods}
\title{The available filtering methods.}
\usage{
filtering_methods()
}
\value{
A character vector of available filtering methods.
}
\description{
The available filtering methods.
}
