% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation}
\alias{imputation}
\title{Impute expression data set with different methods.}
\usage{
imputation(p_df, method, mar_method, mnar_method, ...)
}
\arguments{
\item{p_df}{The data to run filtering on. Of class \code{proteomics_data}.}

\item{method}{The method used for imputation. See \code{imputation_methods()} for
available options.}

\item{mar_method}{The method used for the imputation of missing-at-random
(MAR) missing value if a mixed imputation approach has been chosen. See
\code{imputation_methods_mar()} for available options.}

\item{mnar_method}{The method used for the imputation of missing-not-at-
random (MNAR) missing values if a mixed imputation approach has been
chosen. See \code{imputation_methods_mnar()} for available options.}

\item{...}{Futher arguments passed down to helper functions.}
}
\value{
An object of class \code{proteomics_data} with imputed values.
}
\description{
Impute expression data set with different methods.
}
