% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation_methods}
\alias{imputation_methods}
\alias{imputation_methods_mixed}
\alias{imputation_methods_mar}
\alias{imputation_methods_mnar}
\title{Imputation methods}
\usage{
imputation_methods()

imputation_methods_mixed()

imputation_methods_mar()

imputation_methods_mnar()
}
\description{
\code{imputation_methods()} returns the names of all available imputation methods.

\code{imputation_methods_mixed} returns the names of all available imputation
methods that use mixed imputation.

When a mixed imputation strategy is used, the methods returned by
\code{imputation_methods_mar} and \code{imputation_methods_mnar} are available for
"missing at random" (MAR) imputation and "missing not at random" (MNAR)
imputation, respectively.
}
