% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_obj.R
\name{join_annotation}
\alias{join_annotation}
\title{Join proteomics data with associated annotation.}
\usage{
join_annotation(p_df)
}
\arguments{
\item{p_df}{Proteomics data object.}
}
\value{
Data frame of proteomics data joined with annotation.
}
\description{
Join proteomics data with associated annotation.
}
