% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_obj.R
\name{new_proteomics_data}
\alias{new_proteomics_data}
\title{The OOP framework for storing data in this analysis pipeline.}
\usage{
new_proteomics_data(df, annotation, has_tech_repl)
}
\arguments{
\item{df}{The data frame containing the proteomics data.}

\item{annotation}{A valid annotation data frame.}

\item{has_tech_repl}{A boolean value indicating whether the data has been
log2 transformed.}
}
\value{
A \code{proteomics_data} object.
}
\description{
Contains constructors for and manipulations of the \code{proteomics_data} class.
Low level constructor for \code{proteomics_data} class
}
