% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{stat_analysis_methods}
\alias{stat_analysis_methods}
\title{The available methods for statistical analysis.}
\usage{
stat_analysis_methods()
}
\value{
A character vector with the available methods for statistical
analysis
}
\description{
The available methods for statistical analysis.
}
