% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_obj.R
\name{validate_proteomics_data}
\alias{validate_proteomics_data}
\title{Validator for \code{proteomics_data} class}
\usage{
validate_proteomics_data(proteomics_data)
}
\arguments{
\item{A}{\code{proteomics_data} object.}
}
\value{
A validated \code{proteomics_data} object.
}
\description{
Validator for \code{proteomics_data} class
}
