% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vis_qc_count}
\alias{vis_qc_count}
\title{Plot counts for quality control of proteomics data.}
\usage{
vis_qc_count(p_df)
}
\arguments{
\item{p_df}{Proteomics data.}
}
\description{
Plot the counts of proteins for each sample as a bar chart.
}
