% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vis_qc_histo}
\alias{vis_qc_histo}
\title{Histograms for quality control of proteomics data.}
\usage{
vis_qc_histo(p_df)
}
\arguments{
\item{p_df}{Proteomics data.}
}
\description{
Plot the distributions of the data for each sample as a histogram.
}
