% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vis_qc_scatter}
\alias{vis_qc_scatter}
\title{Scatter plot for proteomics data quality control.}
\usage{
vis_qc_scatter(p_df)
}
\arguments{
\item{p_df}{Proteomics data.}
}
\description{
Plots scatter plots of all pairwise comparisons between two samples in the data set.
}
