% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vis_upset}
\alias{vis_upset}
\title{Plot upset plot of proteomics data.}
\usage{
vis_upset(p_df)
}
\arguments{
\item{p_df}{Proteomics data.}
}
\description{
Visualize which proteins can be found in which groups using an upset plot.
}
