% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{map_uniprot_hgnc}
\alias{map_uniprot_hgnc}
\title{Map UNIPROT IDs to HGNC gene names}
\usage{
map_uniprot_hgnc(
  uniprot_ids,
  hgnc_db,
  multiple_mode = "first",
  manual_file,
  separator_in = ";",
  separator_out = ";"
)
}
\arguments{
\item{uniprot_ids}{A vector of UNIPROT IDs.}

\item{hgnc_db}{A data.frame of the HGNC data base. For this function,
expected to contain the columns "hgnc_id", "symbol", and "uniprot_ids".}

\item{multiple_mode}{Determines how cases are treated where one UNIPROT ID
corresponds to multiple HGNC gene names. Possible values are "first" and
"manual" and "none" (default "first").}

\item{manual_file}{In multiple_mode "manual", the file used for manual
replacements. Required data frame needs to be saved in in a tab separated
file with the columns "input" and "output".}

\item{separator_in}{The separator used to split multiple UNIPROT ID entries
per input. Default is ";".}

\item{separator_out}{The separator used for separating multiple HGNC entries
associated with a single UNIPROT ID, if multiple_mode is "none". The
default is ";".}
}
\value{
Vector of HGNC gene names
}
\description{
Given a vector of UNIPROT IDs , extract corresponding HGNC gene
names where applicable.
}
